+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [TBD]
 ### Changed
-* circus-train-gcp: `credential-provider` can be provided as absolute or relative path to the Google Cloud key file. Using relative path avoids temporary copy of key file to `user.dir`.
+* circus-train-gcp: avoid temporary copy of key file to `user.dir` when using absolute path to Google Cloud credentials file.
 
 ## 12.0.0 - 2018-07-13
 ### Changed